+++ CHANGELOG.md
@@ -16,6 +16,8 @@ now in match expr, only Regex expr was matched as regex, string expr keeps what 
 - add raw mode for type `DateTime`:
 now user could write time like `t'2025-7-10'`
 
+- add `Filesize` module.
+
 ## [0.6.7]
 - eval args for builtin in assign mode
 so user could call `String.red((ls))` and got the expected result.